rename ViewSelector
