+++ src/react/components/ViewSelectorComponent.tsx
